rt: remove runtime::context module
The current runtime thread-local is moved to `runtime::scheduler` as
well as methods to enter and access the current handle.
